python-requirements.txt
# Copyright 2020 ETH Zurich and University of Bologna.
# Licensed under the Apache License, Version 2.0, see LICENSE for details.
# SPDX-License-Identifier: Apache-2.0

bin2coe
dataclasses
editorconfig-checker==2.3.51
flake8
gitpython
hjson
jsonref
jsonschema
mako
progressbar2
tabulate
yamllint
pyyaml
pytablewriter
termcolor
pandas
pyelftools

-r docs/requirements.txt
-r sw/dnn/requirements.txt
